feat: drift detection and takeover implementation (1/ contd)  (#1010)
