+++ pkg/controllers/workapplier/drift_detection_takeover_test.go
@@ -7,7 +7,6 @@ package workapplier
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -19,7 +18,6 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/dynamic/fake"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -451,119 +449,6 @@ func TestPreparePatchDetails(t *testing.T) {
 
 			if diff := cmp.Diff(patchDetails, tc.wantPatchDetails, cmpopts.SortSlices(lessFuncPatchDetail)); diff != "" {
 				t.Fatalf("patchDetails mismatches (-got, +want):\n%s", diff)
-			}
-		})
-	}
-}
-
-// TestDiscardFieldsIrrelevantInComparisonFrom tests the discardFieldsIrrelevantInComparisonFrom function.
-func TestDiscardFieldsIrrelevantInComparisonFrom(t *testing.T) {
-	// This test spec uses a Deployment object as the target.
-	generateName := "app-"
-	dummyManager := "dummy-manager"
-	now := metav1.Now()
-	dummyGeneration := int64(1)
-	dummyResourceVersion := "abc"
-	dummySelfLink := "self-link"
-	dummyUID := "123-xyz-abcd"
-
-	deploy := &appsv1.Deployment{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Deployment",
-			APIVersion: "apps/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:         deployName,
-			Namespace:    nsName,
-			GenerateName: generateName,
-			Annotations: map[string]string{
-				fmt.Sprintf("%s/%s", k8sReservedLabelAnnotationFullDomain, dummyLabelKey): dummyLabelValue1,
-				fmt.Sprintf("%s/%s", k8sReservedLabelAnnotationAbbrDomain, dummyLabelKey): dummyLabelValue1,
-				fmt.Sprintf("%s/%s", fleetReservedLabelAnnotationDomain, dummyLabelKey):   dummyLabelValue1,
-				dummyLabelKey: dummyLabelValue1,
-			},
-			Labels: map[string]string{
-				fmt.Sprintf("%s/%s", k8sReservedLabelAnnotationFullDomain, dummyLabelKey): dummyLabelValue1,
-				fmt.Sprintf("%s/%s", k8sReservedLabelAnnotationAbbrDomain, dummyLabelKey): dummyLabelValue1,
-				fmt.Sprintf("%s/%s", fleetReservedLabelAnnotationDomain, dummyLabelKey):   dummyLabelValue1,
-				dummyLabelKey: dummyLabelValue2,
-			},
-			Finalizers: []string{
-				dummyLabelKey,
-			},
-			ManagedFields: []metav1.ManagedFieldsEntry{
-				{
-					Manager:    dummyManager,
-					Operation:  metav1.ManagedFieldsOperationUpdate,
-					APIVersion: "v1",
-					Time:       &now,
-					FieldsType: "FieldsV1",
-					FieldsV1:   &metav1.FieldsV1{},
-				},
-			},
-			OwnerReferences: []metav1.OwnerReference{
-				dummyOwnerRef,
-			},
-			CreationTimestamp:          now,
-			DeletionTimestamp:          &now,
-			DeletionGracePeriodSeconds: ptr.To(int64(30)),
-			Generation:                 dummyGeneration,
-			ResourceVersion:            dummyResourceVersion,
-			SelfLink:                   dummySelfLink,
-			UID:                        types.UID(dummyUID),
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: ptr.To(int32(1)),
-		},
-		Status: appsv1.DeploymentStatus{
-			Replicas: 1,
-		},
-	}
-	wantDeploy := &appsv1.Deployment{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Deployment",
-			APIVersion: "apps/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Annotations: map[string]string{
-				dummyLabelKey: dummyLabelValue1,
-			},
-			Labels: map[string]string{
-				dummyLabelKey: dummyLabelValue2,
-			},
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: ptr.To(int32(1)),
-		},
-		Status: appsv1.DeploymentStatus{},
-	}
-
-	testCases := []struct {
-		name                string
-		unstructuredObj     *unstructured.Unstructured
-		wantUnstructuredObj *unstructured.Unstructured
-	}{
-		{
-			name:                "deploy",
-			unstructuredObj:     toUnstructured(t, deploy),
-			wantUnstructuredObj: toUnstructured(t, wantDeploy),
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := discardFieldsIrrelevantInComparisonFrom(tc.unstructuredObj)
-
-			// There are certain fields that need to be set manually as they might got omitted
-			// when being cast to an Unstructured object.
-			tc.wantUnstructuredObj.SetFinalizers([]string{})
-			tc.wantUnstructuredObj.SetCreationTimestamp(metav1.Time{})
-			tc.wantUnstructuredObj.SetManagedFields([]metav1.ManagedFieldsEntry{})
-			tc.wantUnstructuredObj.SetOwnerReferences([]metav1.OwnerReference{})
-			unstructured.RemoveNestedField(tc.wantUnstructuredObj.Object, "status")
-
-			if diff := cmp.Diff(got, tc.wantUnstructuredObj, cmpopts.EquateEmpty()); diff != "" {
-				t.Errorf("discardFieldsIrrelevantInComparisonFrom() mismatches (-got, +want):\n%s", diff)
 			}
 		})
 	}
